mm: Move shmem code to tmpfs
This allows a change from anon.cpp to anon.c
